refactor(circuit): split charge-basis operators out of circuit_utils [Clean Code: SRP]
Continuation of the circuit_utils.py junk-drawer split. Move to a new
`scqubits/core/charge_basis_operators.py`:

- `_identity_theta`, `_n_theta_operator`
- `_exp_i_theta_operator`, `_exp_i_theta_operator_conjugate`
- `_cos_theta`, `_sin_theta`

Re-export from circuit_utils.py to preserve existing import paths.
Also corrected two stale rST docstring claims on
`_exp_i_theta_operator` and `_exp_i_theta_operator_conjugate` — both
said "Operator cos(theta)" while implementing e^{±i theta}.
